[APP-68] 리팩토링 (#18)
