+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.GFSample"
+        android:theme="@style/Theme.SeedDesignSample"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.GFSample">
+            android:theme="@style/Theme.SeedDesignSample">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
