+++ component-compose/build.gradle
@@ -28,7 +28,8 @@ android {
         jvmTarget = "1.8"
         freeCompilerArgs = [
                 "-opt-in=kotlin.RequiresOptIn",
-                "-opt-in=androidx.compose.material.ExperimentalMaterialApi"
+                "-opt-in=androidx.compose.material.ExperimentalMaterialApi",
+                "-opt-in=androidx.compose.foundation.ExperimentalFoundationApi",
         ]
     }
     buildFeatures {
