+++ component-compose/src/main/java/com/greenlabsfin/design/component/SeedBottomSheetLayout.kt
@@ -4,29 +4,46 @@ import androidx.compose.animation.core.AnimationSpec
 import androidx.compose.animation.core.TweenSpec
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.gestures.Orientation
 import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.ModalBottomSheetLayout
 import androidx.compose.material.ModalBottomSheetState
 import androidx.compose.material.ModalBottomSheetValue
+import androidx.compose.material.Surface
 import androidx.compose.material.SwipeableDefaults
+import androidx.compose.material.swipeable
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.graphics.isSpecified
 import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.semantics.contentDescription
+import androidx.compose.ui.semantics.dismiss
 import androidx.compose.ui.semantics.onClick
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import com.greenlabsfin.design.core.SeedTheme
 import com.greenlabsfin.design.core.color.black
+import kotlinx.coroutines.launch
+import kotlin.math.roundToInt
 
 typealias SeedBottomSheetState = ModalBottomSheetState
 typealias SeedBottomSheetValue = ModalBottomSheetValue
@@ -41,7 +58,7 @@ fun SeedBottomSheetLayout(
         topStart = 20.dp,
         topEnd = 20.dp,
     ),
-    sheetElevation: Dp = SeedBottomSheetDefaults.Elevation,
+    sheetElevation: Dp = SeedBottomSheetDefaults.elevation,
     sheetBackgroundColor: Color = SeedTheme.colorScheme.container.background,
     sheetContentColor: Color = SeedTheme.colorScheme.contents.neutralPrimary,
     sheetContent: @Composable ColumnScope.() -> Unit,
@@ -75,7 +92,7 @@ fun SeedBottomSheetLayout(
 }
 
 object SeedBottomSheetDefaults {
-    val Elevation = 16.dp
+    val elevation = 16.dp
 
     val scrimColor: Color
         @Composable
@@ -92,7 +109,7 @@ fun rememberSeedBottomSheetState(
         animationSpec,
         saver = SeedBottomSheetState.Saver(
             animationSpec = animationSpec,
-            skipHalfExpanded = true,
+            skipHalfExpanded = false,
             confirmStateChange = confirmStateChange
         )
     ) {
@@ -104,9 +121,88 @@ fun rememberSeedBottomSheetState(
     }
 }
 
+@Composable
+private fun SeedFixedBottomSheetLayout(
+    modifier: Modifier = Modifier,
+    scrimColor: Color = SeedBottomSheetDefaults.scrimColor,
+    sheetState: SeedBottomSheetState = rememberSeedBottomSheetState(SeedBottomSheetValue.Hidden),
+    sheetShape: Shape = RoundedCornerShape(topStart = 20.dp, topEnd = 20.dp),
+    sheetElevation: Dp = SeedBottomSheetDefaults.elevation,
+    sheetBackgroundColor: Color = SeedTheme.colorScheme.container.background,
+    sheetContentColor: Color = SeedTheme.colorScheme.contents.neutralPrimary,
+    sheetContent: @Composable ColumnScope.() -> Unit,
+    content: @Composable () -> Unit,
+) {
+    val scope = rememberCoroutineScope()
+    BoxWithConstraints(modifier) {
+        val fullHeight = constraints.maxHeight.toFloat()
+        val sheetHeightState = remember { mutableStateOf<Float?>(null) }
+        Box(modifier = Modifier.fillMaxSize()) {
+            content()
+            Scrim(
+                color = scrimColor,
+                onDismiss = { scope.launch { sheetState.hide() } },
+                visible = sheetState.targetValue != SeedBottomSheetValue.Hidden
+            )
+        }
+        Surface(
+            modifier = Modifier
+                .fillMaxWidth()
+                .offset {
+                    val y = sheetState.offset.value.roundToInt()
+                    IntOffset(0, y)
+                }
+                .bottomSheetFixedSwipeable(sheetState, fullHeight, sheetHeightState)
+                .onGloballyPositioned {
+                    sheetHeightState.value = it.size.height.toFloat()
+                }
+                .semantics {
+                    if (sheetState.isVisible) {
+                        dismiss {
+                            scope.launch { sheetState.hide() }
+                            true
+                        }
+                    }
+                },
+            shape = sheetShape,
+            elevation = sheetElevation,
+            color = sheetBackgroundColor,
+            contentColor = sheetContentColor,
+        ) {
+            Column(content = sheetContent)
+        }
+    }
+}
+
+private fun Modifier.bottomSheetFixedSwipeable(
+    sheetState: SeedBottomSheetState,
+    fullHeight: Float,
+    sheetHeightState: State<Float?>,
+): Modifier {
+    val sheetHeight = sheetHeightState.value
+    // TODO sheetHeight > 0f <- 요놈 ModalBottomSheetLayout 에 적용 안되어 있음, ModalBottomSheet 도 따로 맹글어야 할수도..
+    val modifier = if (sheetHeight != null && sheetHeight > 0f) {
+        val anchors = mapOf(
+            fullHeight to SeedBottomSheetValue.Hidden,
+            fullHeight - sheetHeight to SeedBottomSheetValue.Expanded
+        )
+        Modifier.swipeable(
+            state = sheetState,
+            anchors = anchors,
+            orientation = Orientation.Vertical,
+            enabled = false,
+            resistance = null,
+        )
+    } else {
+        Modifier
+    }
+
+    return this.then(modifier)
+}
+
 
 @Composable
-internal fun Scrim(
+private fun Scrim(
     color: Color,
     onDismiss: () -> Unit,
     visible: Boolean,
